factor out into subprotocols/sparse; binding / type conversion
