+++ jolt-core/src/subprotocols/mod.rs
@@ -2,4 +2,5 @@
 
 pub mod batched_commitment;
 pub mod grand_product;
+pub mod sparse;
 pub mod sumcheck;
